requirements.txt
langchain
openai
faiss-cpu
tiktoken
python-dotenv
requests
numpy
pandas
langchain-openai
langchain-core